\name{Fpsn}
\alias{Fpsn}
\title{Fpsn}
\description{Function to run the pDPA algorithm with the L2 loss (it is a wrapper to cghseg)}
\usage{Fpsn(x, Kmax)}
\arguments{
  \item{x}{A vector of double : the signal to be segmented}
  \item{Kmax}{
}
}

\value{return a list with a J.est vector containing the L2 loss and a t.est matrix with the changes of the segmentations in 1 to Kmax}

\author{Guillem Rigaill}





