\name{retour_op}
\alias{retour_op}
\title{retour op}
\description{This function is used by the \code{\link{Fpop}} function to recover the best
segment ends from 1:n from the C output.}
\usage{retour_op(path)}
\arguments{
  \item{path}{the \code{path} vector of the "colibri_op_R_c C" function}
}

\value{a vector with the best segment ends.}

\author{Guillem Rigaill <guillem.rigaill@inra.fr> [aut, cre], Toby Hocking [aut], Robert Maidstone [aut], Michel Koskas [ctb], Paul Fearnhead [aut]}





